### 2024-07-24 #7032 @editorialbot
[PRE REVIEW]: TorchSurv: A Lightweight Package for Deep Survival Analysis
**Submitting author:** @melodiemonod (<a href="http://orcid.org/0000-0001-6448-2051">Mélodie Monod</a>)
**Repository:** https://github.com/Novartis/torchsurv
**Branch with paper.md** (empty if default branch): 45-joss-submission
**Version:** v0.1.2
