### 2024-07-25 #7032 @crvernon
:wave: @melodiemonod - Thanks for your submission to JOSS.  While I am getting you set up with a topic editor, please see if you can fix the PDF compile issue flagged in a comment above due to their being an error in your paper or bib file.  Also, please reduce the length of your paper to right around 1000 words and try to fix as many of the missing DOI listed above as you can.  Thanks!
