### 2024-07-26 bbfde39
fix proof of auc probabilistic intepretation
