+++ paper/paper.bib
@@ -3,6 +3,7 @@
   author={Davidson-Pilon, Cameron},
   journal={Journal of Open Source Software},
   volume={4},
+  doi={https://doi.org/10.21105/joss.01317},
   number={40},
   pages={1317},
   year={2019}
@@ -13,17 +14,21 @@
   author={Therneau, Terry M and Lumley, Thomas},
   journal={R Top Doc},
   volume={128},
+  doi={10.32614/CRAN.package.survival},
   number={10},
   pages={28--33},
   year={2015}
 }
 
-@article{paszke2019pytorch,
-  title={Pytorch: An imperative style, high-performance deep learning library},
-  author={Paszke, Adam and Gross, Sam and Massa, Francisco and Lerer, Adam and Bradbury, James and Chanan, Gregory and Killeen, Trevor and Lin, Zeming and Gimelshein, Natalia and Antiga, Luca and others},
-  journal={Advances in neural information processing systems},
-  volume={32},
-  year={2019}
+@misc{paszke2019pytorch,
+      title={PyTorch: An Imperative Style, High-Performance Deep Learning Library}, 
+      author={Adam Paszke and Sam Gross and Francisco Massa and Adam Lerer and James Bradbury and Gregory Chanan and Trevor Killeen and Zeming Lin and Natalia Gimelshein and Luca Antiga and Alban Desmaison and Andreas Köpf and Edward Yang and Zach DeVito and Martin Raison and Alykhan Tejani and Sasank Chilamkurthy and Benoit Steiner and Lu Fang and Junjie Bai and Soumith Chintala},
+      year={2019},
+      eprint={1912.01703},
+      archivePrefix={arXiv},
+      primaryClass={cs.LG},
+      doi={https://doi.org/10.48550/arXiv.1912.01703},
+      url={https://arxiv.org/abs/1912.01703}, 
 }
 
 @article{polsterl2020scikit,
@@ -42,6 +47,7 @@
   author={Potapov, Sergej and Adler, Werner and Schmid, Matthias and Potapov, Maintainer Sergej},
   journal={Statistics in Medicine},
   volume={25},
+  doi={10.32614/CRAN.package.survAUC},
   pages={3474--3486},
   year={2023}
 }
@@ -50,6 +56,7 @@
   title={Momentum contrast for unsupervised visual representation learning},
   author={He, Kaiming and Fan, Haoqi and Wu, Yuxin and Xie, Saining and Girshick, Ross},
   booktitle={Proceedings of the IEEE/CVF conference on computer vision and pattern recognition},
+  doi={10.1109/cvpr42600.2020.00975},
   pages={9729--9738},
   year={2020}
 }
@@ -60,6 +67,7 @@
   journal={BMC medical research methodology},
   volume={18},
   number={1},
+  doi={10.1186/s12874-018-0482-1},
   pages={1--12},
   year={2018},
   publisher={BioMed Central}
@@ -82,6 +90,7 @@
   booktitle={Machine Learning for Healthcare Conference},
   pages={585--608},
   year={2022},
+  doi={https://doi.org/10.48550/arXiv.2204.07276},
   organization={PMLR}
 }
 
@@ -94,6 +103,7 @@
   publisher = {Wiley},
   author = {Cox,  D. R.},
   year = {1972},
+  doi={10.1007/978-1-4612-4380-9_37},
   month = jan,
   pages = {187–202}
 }
@@ -139,6 +149,7 @@
     author = {Sachinthaka Abeywardana},
     year = {2021},
     url = {https://sachinruk.github.io/torchlife//index.html},
+    doi={10.32614/CRAN.package.survival},
   }
 
 @Manual{survAUCpackage,
[…]
