### 2024-07-29 #7032 @tcoroller
Checking if updated DOIs are working:
