### 2024-07-29 842828e
new DOIs, trimmed text
