+++ paper/paper.md
@@ -28,33 +28,18 @@ authors:
     index: 2
   - name: Novartis Pharmaceuticals Corporation, NJ, USA
     index: 3
-date: 18 July 2024
+date: 29 July 2024
 bibliography: paper.bib
 ---
 
 # Summary
 
-`TorchSurv` is a Python [package](https://pypi.org/project/torchsurv/) that serves as a companion tool to perform deep survival modeling within the `PyTorch` environment [@paszke2019pytorch]. With its lightweight design, minimal input requirements, full `PyTorch` backend, and freedom from restrictive parameterizations, `TorchSurv` facilitates efficient deep survival model implementation and is particularly beneficial for high-dimensional and complex data scenarios. At its core, `TorchSurv` features `PyTorch`-based calculations of log-likelihoods for prominent survival models, including the Cox proportional hazards model [@Cox1972] and the Weibull Accelerated Time Failure (AFT) model [@Carroll2003].
-`TorchSurv` has been rigorously tested using both open-source and synthetically generated survival data. The package is thoroughly documented and includes illustrative examples. The latest documentation for TorchSurv can be found on our [website](https://opensource.nibr.com/torchsurv/).
-
-In survival analysis, each observation is associated with a survival reponse, denoted by $y$ (comprising the event indicator and the time-to-event or censoring), and covariates, denoted by $x$. A survival model is parametrized by parameters $\theta$. Within the `TorchSurv` framework, a `PyTorch`-based neural network is defined to act as a flexible function that takes the covariates $x$ as input and outputs the parameters $\theta$. Estimation of the parameters $\theta$ is achieved via maximum likelihood estimation. 
-Additionally, `TorchSurv` offers evaluation metrics, including the time-dependent Area Under under the Receiver operating characteristic (ROC) curve (AUC), the Concordance index (C-index) and the Brier Score, to characterize the predictive performance of survival models.
-Below is an overview of the workflow for model inference and evaluation with `TorchSurv`:
-
-1. Initialize a `PyTorch`-based neural network that defines the function from the covariates $x$ to the parameters $\theta$. In the context of the Cox proportional hazards model for example, the parameters are the log relative hazards.
-2. Initiate training: For each epoch on the training set,
-    -  Draw survival response $y^{\text{train}}$ and covariates $x^{\text{train}}$ from the training set.
-    - Obtain parameters $\theta^{\text{train}}$ given covariates $x^{\text{train}}$ using the neural network.
-    - Calculate the loss as the negative log-likelihood of survival response $y^{\text{train}}$ given parameters $\theta^{\text{train}}$. This calculation is facilitated by `TorchSurv`'s loss function. 
-    - Utilize backpropagation to update the neural network's parameters.
-3. Obtain parameters $\theta^{\text{test}}$ given covariates from the test set $x^{\text{test}}$ using the trained neural network.
-4. Evaluate the predictive performance of the model using `TorchSurv`'s evaluation metric functions given parameters $\theta^{\text{test}}$ and survival response from the test set $y^{\text{test}}$.
-
-
[…]
