+++ paper/paper.bib
@@ -3,7 +3,7 @@
   author={Davidson-Pilon, Cameron},
   journal={Journal of Open Source Software},
   volume={4},
-  doi={https://doi.org/10.21105/joss.01317},
+  doi={10.21105/joss.01317},
   number={40},
   pages={1317},
   year={2019}
@@ -27,7 +27,7 @@
       eprint={1912.01703},
       archivePrefix={arXiv},
       primaryClass={cs.LG},
-      doi={https://doi.org/10.48550/arXiv.1912.01703},
+      doi={10.48550/arXiv.1912.01703},
       url={https://arxiv.org/abs/1912.01703}, 
 }
 
@@ -39,7 +39,8 @@
   number={1},
   pages={8747--8752},
   year={2020},
-  publisher={JMLRORG}
+  publisher={JMLRORG},
+  doi={10.5281/zenodo.3352342}
 }
 
 @article{potapov2023packagesurvAUC,
@@ -90,7 +91,7 @@
   booktitle={Machine Learning for Healthcare Conference},
   pages={585--608},
   year={2022},
-  doi={https://doi.org/10.48550/arXiv.2204.07276},
+  doi={10.48550/arXiv.2204.07276},
   organization={PMLR}
 }
 
@@ -142,6 +143,7 @@
     year = {2024},
     note = {R package version 3.5-8},
     url = {https://CRAN.R-project.org/package=survival},
+    doi={10.32614/CRAN.package.survival},
   }
 
 @Manual{torchlifeAbeywardana,
