### 2024-07-29 #7032 @tcoroller
> **Paper file info**:
>
> 📄 Wordcount for `paper.md` is **1372**
>
> ✅ The paper includes a `Statement of need` section

Hi @crvernon , thank you for helping us prepare the review. I will take over the process while Melodie is away.

- I have fixed the DOIs - with two that I just cannot find
- I have trimmed the text to 1040 words, excluding the markdown tags (authors, affiliations) and the two table links. Please let me know if thats acceptable to you now
