### 2024-07-29 #7032 @tcoroller
> ⚠️ [An error happened when generating the pdf](https://github.com/openjournals/joss-papers/actions/runs/10146354788?check_suite_focus=true).

I am unsure how to fix this one. I don't spot anything wrong in the markdown and the log error is unclear to me. Would anyone be able to provide more guidance on this?

``` shell
Run bundle exec ruby /home/runner/work/_actions/xuanxu/paper-action/main/get_paper.rb
/home/runner/work/_actions/xuanxu/paper-action/main/vendor/bundle/ruby/3.3.0/gems/theoj-1.10.0/lib/theoj/paper.rb:141:in `parse_affiliations': undefined method `each' for nil (NoMethodError)

        affiliations_yaml.each do |affiliation|
                         ^^^^^
	from /home/runner/work/_actions/xuanxu/paper-action/main/vendor/bundle/ruby/3.3.0/gems/theoj-1.10.0/lib/theoj/paper.rb:120:in `parse_authors'
	from /home/runner/work/_actions/xuanxu/paper-action/main/vendor/bundle/ruby/3.3.0/gems/theoj-1.10.0/lib/theoj/paper.rb:24:in `authors'
	from /home/runner/work/_actions/xuanxu/paper-action/main/vendor/bundle/ruby/3.3.0/gems/theoj-1.10.0/lib/theoj/submission.rb:103:in `all_metadata'
	from /home/runner/work/_actions/xuanxu/paper-action/main/vendor/bundle/ruby/3.3.0/gems/theoj-1.10.0/lib/theoj/submission.rb:[78](https://github.com/openjournals/joss-papers/actions/runs/10146354788/job/28054190314#step:3:86):in `metadata_info'
	from /home/runner/work/_actions/xuanxu/paper-action/main/vendor/bundle/ruby/3.3.0/gems/theoj-1.10.0/lib/theoj/submission.rb:57:in `article_metadata'
	from /home/runner/work/_actions/xuanxu/paper-action/main/get_paper.rb:25:in `<main>'

```
