### 2024-07-29 42a95a5
safer eval model
