### 2024-07-31 ba1f542
added badge conda
