### 2024-07-31 41f5911
Added conda information
