+++ _sources/CHANGELOG.md
@@ -1,0 +1,22 @@
+Change log
+=========
+
+Version 0.1.2
+-------------
+
+* Updated package documentation with publication links & badges (#9, #14, #16, #19, #21, #22, #24)
+* Fixed and documented package dependencies (#1)
+
+Version 0.1.1
+-------------
+
+* Added `metrics` classes (AUC, Cindex, Brier score)
+* Added `stats` class with Kaplan Meier
+* Created Sphinx documentation
+* Created `notebook` examples
+* Added R benchmark comparison
+
+Version 0.1.0
+-------------
+
+* Initial release of CoxPH and Weibull classes.
