### 2024-08-09 015e45d
clamped log_hazard to prevent torch.Inf
