+++ docs/CHANGELOG.md
@@ -1,6 +1,13 @@
 Change log
 =========
 
+Version 0.1.3
+-------------
+
+* Tutorial dataset error on momentum.ipynb #50
+* Fix issue #48 - log_hazard returns torch.Inf
+* Updated CICD builds
+
 Version 0.1.2
 -------------
 
