+++ .github/workflows/codeqc.yml
@@ -39,7 +39,7 @@ jobs:
 
     - name: Update conda environment
       run:
-        conda env update -n torchsurv -f dev/environment.yml
+        conda env update -n torchsurv -c conda-forge -f dev/environment.yml
       if: steps.cache.outputs.cache-hit != 'true'
 
     - name: Code QC
@@ -47,7 +47,7 @@ jobs:
       run: |
         conda activate torchsurv
         ./dev/codeqc.sh
- 
+
     - name: Tests
       shell: bash -l {0}
       run: |
