### 2024-09-26 71ffc30
make formatting mandatory & reformat kaplan meier test
