+++ docs/CHANGELOG.md
@@ -6,7 +6,10 @@ Version 0.1.3 (unreleased)
 
 * Tutorial dataset error on momentum.ipynb #50
 * Fix issue #48 - log_hazard returns torch.Inf
-* Updated CICD builds
+* Fix warning with Spearman correlation #41
+* Added in-depth statistical background to link AUC to C-index #39
+* Created Conda Forge version #47
+* Updated CICD builds #53
 
 Version 0.1.2
 -------------
