### 2024-09-26 0fc78c3
adding sphinx-issues for Changelog links
