+++ dev/environment.yml
@@ -26,4 +26,3 @@ dependencies:
   - twine=4.0.2
   - scipy=1.12.0
   - sphinx-math-dollar=1.2.1
-  - sphinx-issues=4.1.0
