### 2024-09-26 4e4d8fb
try fix CI builds (#53)
* try fix builds

* partial rollback

* fix build dependency

* make formatting mandatory & reformat kaplan meier test

* add unreleased note

* edited

* adding sphinx-issues for Changelog links

* reverted change

---------

Co-authored-by: corolth1 <thibaud.coroller@novartis.com>
