+++ tests/test_kaplan_meier.py
@@ -208,9 +208,9 @@ class TestNonParametric(unittest.TestCase):
         for batch in batch_container.batches:
             (train_time, train_event, test_time, *_) = batch
 
-            train_event[
-                -1
-            ] = False  # if last event is censoring, the last KM is > 0 and it cannot predict beyond this time
+            train_event[-1] = (
+                False  # if last event is censoring, the last KM is > 0 and it cannot predict beyond this time
+            )
             km = KaplanMeierEstimator()
             km(train_event, train_time, censoring_dist=False)
 
