+++ .github/workflows/build-release.yml
@@ -27,7 +27,7 @@ jobs:
         run: python -m twine check dist/*
 
       - name: Upload artifact
-        uses: actions/upload-artifact@v2
+        uses: actions/upload-artifact@v4
         with:
           name: Python-package
           path: dist
