### 2024-10-03 #7032 @kanishkan91
@WeakCha given your experience with JOSS and your recent reviews, would you be interested in reviewing this paper? I know you just got done with a review 3 weeks ago. Let me know what you think.
