### 2024-10-03 #7032 @kanishkan91
@RhysPeploe would you be interested in reviewing this paper given your previous experience with JOSS and interests? It seemed like a good fit.
