### 2024-10-03 #7032 @kanishkan91
@mhu48 Given your previous experience reviewing for JOSS, would you be interested in reviewing this paper? It seems like a good fit given your interests and profile. Let me know what you think.
