### 2024-10-07 86de946
Update devnotes.md
Add instructions for conda-forge releases.
