+++ _sources/benchmarks.rst
@@ -1,0 +1,2 @@
+.. include:: ../tests/README.md
+    :parser: myst_parser.sphinx_
