+++ _sources/README-copy.md
@@ -1,0 +1,251 @@
+# Deep survival analysis made easy
+
+![CodeQC](https://github.com/Novartis/torchsurv/actions/workflows/codeqc.yml/badge.svg?branch=main)
+![Docs](https://github.com/Novartis/torchsurv/actions/workflows/docs.yml/badge.svg?branch=main)
+[![PyPI - Version](https://img.shields.io/pypi/v/torchsurv?)](https://pypi.org/project/torchsurv/)
+[![Conda](https://img.shields.io/conda/v/conda-forge/torchsurv?label=conda)](https://anaconda.org/conda-forge/torchsurv)
+[![arXiv](https://img.shields.io/badge/arXiv-2404.10761-f9f107.svg?)](https://arxiv.org/abs/2404.10761)
+[![Documentation](https://img.shields.io/badge/GithubPage-Sphinx-blue)](https://opensource.nibr.com/torchsurv/)
+[![PyPI Downloads](https://img.shields.io/pypi/dm/torchsurv.svg?label=PyPI%20downloads)](
+https://pypi.org/project/torchsurv/)
+[![Conda Downloads](https://img.shields.io/conda/dn/conda-forge/torchsurv.svg?label=Conda%20downloads)](
+https://anaconda.org/conda-forge/torchsurv)
+
+`TorchSurv` is a Python package that serves as a companion tool to perform deep survival modeling within the `PyTorch` environment. Unlike existing libraries that impose specific parametric forms on users, `TorchSurv` enables the use of custom `PyTorch`-based deep survival models.  With its lightweight design, minimal input requirements, full `PyTorch` backend, and freedom from restrictive survival model parameterizations, `TorchSurv` facilitates efficient survival model implementation, particularly beneficial for high-dimensional input data scenarios.
+
+# TL;DR
+
+Our idea is to **keep things simple**. You are free to use any model architecture you want! Our code has 100% PyTorch backend and behaves like any other functions (losses or metrics) you may be familiar with.
+
+Our functions are designed to support you, not to make you jump through hoops. Here's a pseudo code illustrating how easy is it to use `TorchSurv` to fit and evaluate a Cox proportional hazards model:
+
+```python
+from torchsurv.loss import cox
+from torchsurv.metrics.cindex import ConcordanceIndex
+
+# Pseudo training loop
+for data in dataloader:
+    x, event, time = data
+    estimate = model(x)  # shape = torch.Size([64, 1]), if batch size is 64
+    loss = cox.neg_partial_log_likelihood(estimate, event, time)
+    loss.backward()  # native torch backend
+
+# You can check model performance using our evaluation metrics, e.g, the concordance index with
+cindex = ConcordanceIndex()
+cindex(estimate, event, time)
+
+# You can obtain the confidence interval of the c-index
+cindex.confidence_interval()
+
+# You can test whether the observed c-index is greater than 0.5 (random estimator)
+cindex.p_value(method="noether", alternative="two_sided")
+
+# You can even compare the metrics between two models (e.g., vs. model B)
+cindex.compare(cindexB)
+```
+
+# Installation and dependencies
+
+
+First, install the package using either [PyPI]([https://pypi.org/](https://pypi.org/project/torchsurv/)) or [Conda]([https://anaconda.org/anaconda/conda](https://anaconda.org/conda-forge/torchsurv))
+
+- Using conda (`recommended`)
+```bash
+conda install conda-forge::torchsurv
+```
+- Using PyPI
+```bash
+pip install torchsurv
+```
+
+- Using for local installation (`latest version`)
+
+```bash
+git clone <repo>
+cd <repo>
+pip install -e .
+```
+
[…]
