+++ _sources/CHANGELOG.md
@@ -1,0 +1,32 @@
+Change log
+=========
+
+Version 0.1.3
+-------------
+
+* Tutorial dataset error on momentum.ipynb #50
+* Fix issue #48 - log_hazard returns torch.Inf
+* Fix warning with Spearman correlation #41
+* Added in-depth statistical background to link AUC to C-index #39
+* Created Conda Forge version #47
+* Updated CICD builds #53
+
+Version 0.1.2
+-------------
+
+* Updated package documentation with publication links & badges (#9, #14, #16, #19, #21, #22, #24)
+* Fixed and documented package dependencies (#1)
+
+Version 0.1.1
+-------------
+
+* Added `metrics` classes (AUC, Cindex, Brier score)
+* Added `stats` class with Kaplan Meier
+* Created Sphinx documentation
+* Created `notebook` examples
+* Added R benchmark comparison
+
+Version 0.1.0
+-------------
+
+* Initial release of CoxPH and Weibull classes.
