### 2024-10-10 #7032 @LingfengLuo0510
Hi Kanishka,

Yes I am interested.

Best,
Lingfeng

On Wed, Oct 9, 2024 at 19:56 Kanishka Narayan ***@***.***>
wrote:

> @LingfengLuo0510 <https://github.com/LingfengLuo0510> would you be
> interested in reviewing this manuscript for JOSS? I see that you just got
> done with another review review a month ago. Let me know what you think.
>
