### 2024-10-10 #7032 @mhu48
Yes, happy to review.

Kanishka Narayan ***@***.***> 于 2024年10月2日周三 21:51写道：

> @mhu48 <https://github.com/mhu48> Given your previous experience
> reviewing for JOSS, would you be interested in reviewing this paper? It
> seems like a good fit given your interests and profile. Let me know what
> you think.
>
