### 2024-10-21 Novartis/torchsurv#58 @melodiemonod
comments JOSS 241018
Address WeakCha's comments from 241018

https://github.com/openjournals/joss-reviews/issues/7341#issuecomment-2423186227

- [x] Adding an example after you introduce the functionality in the paper
- [x] There are some typos in your paper.
- [x]  show what helpers_introduction is
