### 2024-10-21 Novartis/torchsurv#58 @melodiemonod
@tcoroller , I think for action 3, we can just add the functions from helpers_introduction in blocks of the notebook?
