### 2024-10-21 Novartis/torchsurv#58 @tcoroller
@melodiemonod I kinda disagree, especially since the `helpers` are located in the same [folder](https://github.com/Novartis/torchsurv/tree/main/docs/notebooks) as in the notebooks. They are standalone function meant to keep the notebooks lean. There are no TorchSurv related function there but simply boilerplate ones.

Let me reply to the reviewers with those details and maybe highlighting the link to the folder in the text (maybe listing deps if needed) and if this is not suitable we can find another solution. What do you think?
