### 2024-10-21 Novartis/torchsurv#58 @melodiemonod
@tcoroller  please review the example in paper.md
