### 2024-10-21 #7341 @tcoroller
Dear @WeakCha,

Thank you for your early comments. Regarding your last one (`helpers_*`), **would it be ok with you** if we add this line below in the notebooks with the link to the helper file itself?

For the first notebook (`introduction`),  the [helpers_introduction] line will be the following: (https://github.com/Novartis/torchsurv/blob/main/docs/notebooks/helpers_introduction.py)
```python
# PyTorch boilerplate - see https://github.com/Novartis/torchsurv/blob/main/docs/notebooks/helpers_introduction.py
from helpers_introduction import Custom_dataset, plot_losses
```

I would discourage adding the code directly in the notebook because it would overcrowd the notebook:

1. I would need to be at the beginning of the notebook to be properly loaded and run, which is annoying because one helper ([helpers_momentum](https://github.com/Novartis/torchsurv/blob/main/docs/notebooks/helpers_momentum.py)) is ~200lines of code
2. They are simply boilerplate PyTorch code, with no relationship to TorchSurv and well established function (model and datasets, plot losses).

If our suggestion isn't satisfying, I can look into other option (one could be to have the cell hidden, but requires another package thus making it rely on more dependancies)
