### 2024-10-21 Novartis/torchsurv#58 @melodiemonod
(Here https://github.com/Novartis/torchsurv/blob/58-comments-joss-241018/paper/paper.md)
