### 2024-10-21 Novartis/torchsurv#58 @tcoroller
@melodiemonod

Two big comments: I

1. Your branch is behind `main` (like by a lot), so we can either update it or cherry pick the change and do a new PR from current main (its only the `.md` anyway)
2. I like the text updates but I don't understand the example:

Minor questions here:
- All your events are `1`, is that a mistake or by design?
- The fixed values are not explained (e.g., exponantial value, end_time, etc..)
- What is the goal of the weights multiplication for the time?

I tinkered a lot with the example, but could not get results as good as yours. Mostly because you are testing on the trained datasets. Here's the same but now using a generator (sample as needed the distribution)..

```python
import torch
from torch.utils.data import DataLoader, Dataset
from torchsurv.loss import cox
from torchsurv.metrics.cindex import ConcordanceIndex
from torchsurv.metrics.auc import Auc

torch.manual_seed(42)

# 1. Simulate batch data. Ensure that your data are formatted in the same way
n_samples = 1000  # number of observations
batch_size = 32  # batch size

# Define the generator function
def tte_generator(batch_size: int, n_features: int = 10):
    while True:
        x = torch.randn(batch_size, n_features)  # 10 features for each observation
        weights = torch.randn(n_features)  # weights associated with the features
        event = torch.randint(
            low=0, high=2, size=(batch_size,), dtype=torch.bool
        )  # event indicator
        time = torch.abs(800 + x @ weights + torch.randn(batch_size) * 25)  # event time
        yield x, event, time

# 2. Define the PyTorch dataset class
class TTE_dataset(Dataset):
    def __init__(self, generator: callable, batch_size: int):
        self.generator = generator(batch_size=batch_size)

    def __len__(self):
        return n_samples // batch_size  # Number of batches

    def __getitem__(self, index):
        return next(self.generator)

# 3. Define the backbone model on the log hazards.
class MyPyTorchCoxModel(torch.nn.Module):
    def __init__(self):
        super(MyPyTorchCoxModel, self).__init__()
        self.fc = torch.nn.Linear(10, 1, bias=False)  # Simple linear model

    def forward(self, x):
        return self.fc(x)

# 4. Instantiate the model, optimizer, dataset and dataloader
cox_model = MyPyTorchCoxModel()
optimizer = torch.optim.Adam(cox_model.parameters(), lr=0.01)
dataset = TTE_dataset(tte_generator, batch_size=batch_size)
dataloader = DataLoader(
    dataset, batch_size=1, shuffle=True
)  # Batch size of 1 because dataset yields batches

# 5. Training loop
for epoch in range(100):
    for i, batch in enumerate(dataloader):
        x, event, time = [t.squeeze() for t in batch]  # Squeeze extra dimension
        optimizer.zero_grad()
        log_hzs = cox_model(x)  # torch.Size([16, 1])
        loss = cox.neg_partial_log_likelihood(
            log_hzs, event, time, reduction="mean"
        )  # use mean log-likelihood
        loss.backward()
        optimizer.step()

# 6. Evaluate the model
# Generate a new batch of data to evaluate
new_batch = next(tte_generator(batch_size=n_samples))
x, event, time = [t.squeeze() for t in new_batch]
log_hzs = cox_model(x)  # Trained log hazards

# AUC at time point 800
auc = Auc()
print(
    "AUC:", auc(log_hzs, event, time, new_time=torch.tensor(800.0))
)  # tensor([0.5531])
print("AUC Confidence Interval:", auc.confidence_interval())  # tensor([0.5220, 0.5841])
print("AUC p-value:", auc.p_value(alternative="greater"))  # tensor([0.0004])

# C-index
cindex = ConcordanceIndex()
print("C-Index:", cindex(log_hzs, event, time))  # tensor(0.5465)
print(
    "C-Index Confidence Interval:", cindex.confidence_interval()
)  # tensor([0.4897, 0.6033])
print("C-Index p-value:", cindex.p_value(alternative="greater"))  # tensor(0.0544)
```
