### 2024-10-22 Novartis/torchsurv#58 @melodiemonod
Hello @tcoroller!

I just pushed a new update of the paper with the example (Here https://github.com/Novartis/torchsurv/blob/58-comments-joss-241018/paper/paper.md). And also added the example below for reference.

> Your branch is behind main (like by a lot), so we can either update it or cherry pick the change and do a new PR from current main (its only the .md anyway)

I used the JOSS branch as the source for the review branch. So after we are happy with the updates we can merge 58 -> 45. If you want, we can then update 45 with the update from main, 45 <- main?

> All your events are 1, is that a mistake or by design?

You removed from the updated example the true underlying generative mechanism for time-to-event data, i.e., each individual is associated with a time-to-event (`event_time`) and time-to-censoring (`censoring_time`), and only the first of the two (`time`) is observed. It is important to reproduce this mechanism so we can create a correlation between the covariates and the time to event, `event_time` not `time`. We cannot create a correlation with `time` directly as you did there (`time = torch.abs(800 + x @ weights + torch.randn(batch_size) * 25)`) because it would mean that the time of censoring (which time is partly composed of) would also correlated with the covariates.
--> I added back the true mechanism in the generator function. Also, the events were all 1 because I used a mean of time-to-censoring that was too large (I reduced it now).

> What is the goal of the weights multiplication for the time?

The time-to-event (`event_time`) depends linearly on the features weighted by `weights`. You cannot sample from them every time in the generator function, they need to be fixed for all individuals. This is why you did not get good metrics.
--> I put the generation of weights outside of the generator function.

Updated example:
```python
import torch
from torch.utils.data import DataLoader, Dataset
from torchsurv.loss import cox
from torchsurv.metrics.cindex import ConcordanceIndex
from torchsurv.metrics.auc import Auc

torch.manual_seed(42)

# 1. Simulate batch data.
# Ensure that your x, event and time are formatted in the same way
n_features = 10  # int, number of features per observation
time_end = torch.tensor(
    2000.0
)  # float, end of observational period after which all observations are censored
weights = (
    torch.randn(n_features) * 5
)  # float, weights associated with the features ~ normal(0, 5^2)

# Define the generator function
def tte_generator(batch_size: int):
    while True:
        x = torch.randn(batch_size, n_features)  # features

        mean_event_time, mean_censoring_time = 1000.0 + x @ weights, 1000.0

        event_time = (
            mean_event_time + torch.randn(batch_size) * 50
        )  # event time ~ normal(mean_event_time, 50^2)
        censoring_time = torch.distributions.Exponential(
            1 / mean_censoring_time
        ).sample(
            (batch_size,)
        )  # censoring time ~ Exponential(mean = mean_censoring_time)
        censoring_time = torch.minimum(
            censoring_time, time_end
        )  # truncate censoring time to time_end

        event = (event_time <= censoring_time).bool()  # event indicator
        time = torch.minimum(event_time, censoring_time)  # observed time

        yield x, event, time

# 2. Define the PyTorch dataset class
class TTE_dataset(Dataset):
    def __init__(self, generator: callable, batch_size: int):
        self.batch_size = batch_size
        self.generatated_data = generator(batch_size=batch_size)

    def __len__(self):
        return self.batch_size

    def __getitem__(self, index):
        return next(self.generatated_data)

# 3. Define the backbone model on the log hazards.
class MyPyTorchCoxModel(torch.nn.Module):
    def __init__(self):
        super(MyPyTorchCoxModel, self).__init__()
        self.fc = torch.nn.Linear(10, 1, bias=False)  # Simple linear model

    def forward(self, x):
        return self.fc(x)

# 4. Instantiate the model, optimizer, dataset and dataloader
cox_model = MyPyTorchCoxModel()
optimizer = torch.optim.Adam(cox_model.parameters(), lr=0.01)

batch_size = 64  # int, batch size
dataset = TTE_dataset(tte_generator, batch_size=batch_size)
dataloader = DataLoader(
    dataset, batch_size=1, shuffle=True
)  # Batch size of 1 because dataset yields batches

# 5. Training loop
for epoch in range(100):
    for i, batch in enumerate(dataloader):
        x, event, time = [t.squeeze() for t in batch]  # Squeeze extra dimension
        optimizer.zero_grad()
        log_hzs = cox_model(x)  # torch.Size([batch_size, 1])
        loss = cox.neg_partial_log_likelihood(log_hzs, event, time)
        loss.backward()
        optimizer.step()

# 6. Evaluate the model
n_samples_test = 1000  # int, number of observations in test set

data_test = next(tte_generator(batch_size=n_samples_test))
x, event, time = [t.squeeze() for t in data_test]  # test set
log_hzs = cox_model(x)  # log hazards evaluated on test set

# AUC at time point 1000
auc = Auc()
print(
    "AUC:", auc(log_hzs, event, time, new_time=torch.tensor(1000.0))
)  # tensor([0.5902])
print("AUC Confidence Interval:", auc.confidence_interval())  # tensor([0.5623, 0.6180])
print("AUC p-value:", auc.p_value(alternative="greater"))  # tensor([0.])

# C-index
cindex = ConcordanceIndex()
print("C-Index:", cindex(log_hzs, event, time))  # tensor(0.5774)
print(
    "C-Index Confidence Interval:", cindex.confidence_interval()
)  # tensor([0.5086, 0.6463])
print("C-Index p-value:", cindex.p_value(alternative="greater"))  # tensor(0.0138)
```
