### 2024-10-22 Novartis/torchsurv#58 @tcoroller
> I used the JOSS branch as the source for the review branch. So after we are happy with the updates we can merge 58 -> 45. If you want, we can then update 45 with the update from main, 45 <- main?

I wonder if we are better off just making a new PR from main and just carry over the `.md`. Lots of changes.

I can do this piece
