+++ paper/paper.md
@@ -82,7 +82,7 @@ for data in dataloader:
 from torchsurv.loss import weibull
 
 # PyTorch model outputs TWO Weibull parameters per observation
-my_model = MyPyTorcWeibullhModel() 
+my_model = MyPyTorchWeibullhModel() 
 
 for data in dataloader:
     x, event, time = data
@@ -95,7 +95,8 @@ for data in dataloader:
 
 ```python
 from torchsurv.loss import Momentum
-my_model = MyPyTorchXCoxModel()  
+
+my_model = MyPyTorchCoxModel()  
 my_loss = cox.neg_partial_log_likelihood  # Works with any TorchSurv loss
 momentum = Momentum(backbone=my_model, loss=my_loss)
 
@@ -115,7 +116,7 @@ The `TorchSurv` package offers a comprehensive set of metrics to evaluate the pr
 **AUC.** The AUC measures the discriminatory capacity of the survival model at a given time $t$, i.e., the ability to provide a reliable ranking of times-to-event based on estimated subject-specific risk scores [@Heagerty2005;@Uno2007;@Blanche2013].
 
 ```python
-from torchsurv.metrics import Auc
+from torchsurv.metrics.auc import Auc
 auc = Auc()
 auc(log_hzs, event, time)  # AUC at each time 
 auc(log_hzs, event, time, new_time=torch.tensor(10.))  # AUC at time 10
@@ -124,7 +125,7 @@ auc(log_hzs, event, time, new_time=torch.tensor(10.))  # AUC at time 10
 **C-index.** The C-index is a generalization of the AUC that represents the assessment of the discriminatory capacity of the survival model across the entire time period [@Harrell1996;@Uno_2011].
 
 ```python
-from torchsurv.metrics import ConcordanceIndex
+from torchsurv.metrics.cindex import ConcordanceIndex
 cindex = ConcordanceIndex()
 cindex(log_hzs, event, time)
 ```
@@ -132,7 +133,7 @@ cindex(log_hzs, event, time)
 **Brier Score.** The Brier score evaluates the accuracy of a model at a given time $t$ [@Graf_1999]. It represents the average squared distance between the observed survival status and the predicted survival probability. The Brier score cannot be obtained for the Cox proportional hazards model because the survival function is not available, but it can be obtained for the Weibull ATF model.
 
 ```python
-from torchsurv.metrics import Brier
+from torchsurv.metrics.brier_score import BrierScore
 surv = survival_function(log_params, time) 
 brier = Brier()
 brier(surv, event, time)  # Brier score at each time
@@ -146,6 +147,128 @@ cindex.confidence_interval()  # CI, default alpha=.05
 cindex.p_value(alternative='greater')  # pvalue, H0:c=0.5, HA:c>0.5
 cindex.compare(cindex_other)  # pvalue, H0:c1=c2, HA:c1>c2
 ```
+
+# Comprehensive Example: Fitting a Cox Proportional Hazards Model with TorchSurv
+
[…]
