### 2024-10-22 Novartis/torchsurv#58 @tcoroller
@melodiemonod please check #59 to make sure I didn't forgot anything.
