### 2024-10-22 Novartis/torchsurv#58 @melodiemonod
@tcoroller looks all good to me! I just left one comment about a typo (https://github.com/Novartis/torchsurv/pull/59/commits/f6d624c92f91270d90eda4a1de23dbed6b8740ee#r1811249877), do you think you could fix it before merging?

Also, for the JOSS submission, we had to specify the branch to do the review on, which for now was 45-XX. The bot is looking in this branch to generate the paper proof. We need to ask the editor how we can change this "branch to review" to main.
