### 2024-10-23 Novartis/torchsurv#58 @tcoroller
Let me close this comment and merge to main. I will also tell the reviewer that they should be looking at `main`
