### 2024-10-23 d460f9c
JOSS submission (#59)
* edits

* ⭐

* fixed typo

* added boilerplate links
