### 2024-10-23 #7341 @tcoroller
Hello everyone,

We modified the package according to the first comments. Please use the `main` branch (default) for the review as we will keep this one up to date with your comments.

Thanks
