+++ _sources/_autosummary/torchsurv.metrics.brier_score.rst
@@ -1,0 +1,29 @@
+﻿torchsurv.metrics.brier\_score
+==============================
+
+.. automodule:: torchsurv.metrics.brier_score
+
+   
+   
+   
+
+   
+   
+   
+
+   
+   
+   .. rubric:: Classes
+
+   .. autosummary::
+   
+      BrierScore
+   
+   
+
+   
+   
+   
+
+
+
