### 2024-10-24 57845ed
greater OR EQUAL
