### 2024-10-29 ec56ad4
Centered logo
