### 2024-10-31 #7341 @tcoroller
dear @WeakCha ,

Thank you for your comments, please see my replies below:

1. **Learning rate**: You may be referring to the [introduction notebook](https://opensource.nibr.com/torchsurv/notebooks/introduction.html#Section-1.2:-MLP-model-training). Indeed, the Cox model isn't doing well, with a poor looking training loss profile. However, this notebook also shows the Weibull model doing pretty well and then ultimely compare the two to illustrate our functions, such as  this [key function](https://opensource.nibr.com/torchsurv/_autosummary/torchsurv.metrics.cindex.html#torchsurv.metrics.cindex.ConcordanceIndex.compare) `cindex.compare()`. This notebook is meant to illustrate the function rather than pushing for performance, and having Cox model performing poorly was good to this purpose. Please let me know that clarifies and give you better context.
2. **helper_functions**: Sorry for the confusion there, those functions are **not** part of `TorchSurv` but only support the notebooks. Installing `TorchSurv` will not solve this, but those function are modules sitting directly next to the notebooks, so  the import is relative and should run directly. You need to make sure you have all files together, or better clone the whole repository to be safe. Please see below in the screenshot.

3. **MNIST**: Appologies for the trouble running this notebook. I made this example on a M2 macbook with 64Gb of memory, so I guess I didn't think too much of batch size limitation. Several colleagues tried it on HPC, but I would assume with GPUs. I will add a code snippet to downgrade the batch size (currently 500) to 50 or less to make sure it runs smoothly on smaller computer if the user is using CPUs.
4. **Model comparison**: Thank you for this comment, this is a very important part of our work to ensure that our package is comparable against [established R/python packages](https://opensource.nibr.com/torchsurv/index.html#related-packages). During the comparison, we focused on simple model (single neuron, batch size equal to dataset) to mimic other packages and validate the code. We did not compare our method to other deep learning package, simply because most are either deprecated (last commit >3 years ago), poorly documented or simply out of scope. We have an extensive list of comparison for all our metrics that can be found [here](https://opensource.nibr.com/torchsurv/benchmarks.html). @melodiemonod is the lead for this piece and can comment further.
5. **Formatting**: Thank you for the comment and sorry for the inconsistent formatting. We will review and edit the document accordingly.
6. **Typos**: Thank you for the comment and sorry for the typos. We will review and edit the document accordingly.
7. **Paper**: Thank you for raising the issue about the paper not being update. We are now working on the `main` branch but I think the `editorialbot` is still look at an older branch. @kanishkan91  could you help update the `Branch with paper.md` to `main`?

Thank you,
Thibaud
