### 2024-10-31 Novartis/torchsurv#66 @tcoroller
JOSS-review#2
see https://github.com/openjournals/joss-reviews/issues/7341#issuecomment-2449914054

- [ ] Downgrade batch size for MNIST notebook when running only on CPUs (or low RAM)
- [ ] Fix typos (e.g.,  "recommanded" should be "recommended", "this notebooks" should be "this notebook".)
- [ ] Inconsistent display effect such as `here`, **here** or **`here`** to emphasize
