### 2024-10-31 #7341 @melodiemonod
To expand on point 4 about model comparison: our package is designed to provide all the necessary functions to assist in model fitting (i.e.., with the negative log-likelihood function) and evaluate its performance (i.e., with evaluation metric functions), rather than providing functions for directly fitting a model. Unlike packages like `deepsurv`, we do not provide functions like `fit` for direct model training. Instead, our package focuses on offering users maximum flexibility, allowing them to integrate our functions seamlessly within their own PyTorch-based model-fitting workflows. As a result, comparing model-fitting outcomes with another algorithm falls outside the scope of our package. However, we have included a comparison to demonstrate that our likelihood and metric functions produce results consistent with established package. This is the [notebook](https://opensource.nibr.com/torchsurv/benchmarks.html) that Thibaud mentioned above.
