### 2024-11-08 #7341 @tcoroller
Dear @WeakCha,

I have updated accordingly the documentation (see [Pull Request](https://github.com/Novartis/torchsurv/pull/68/files)) with the requested changes.

Thank you!
Thibaud
