### 2024-11-11 d2f66d7
updating (#69)
* Centered logo

* set model to eval model before parameters evaluation on test (#65)

* review #2 (#68)

---------

Co-authored-by: Mélodie Monod <monod.melodie@gmail.com>
