### 2024-11-21 #7341 @kanishkan91
@XinyiEmilyZhang, your name was recommended as a potential reviewer. Would you be interested in reviewing this package for the Journal of Open Source Software (JOSS)?
