### 2024-11-21 #7341 @XinyiEmilyZhang
Thanks for letting me know! I’d like to review the package. Please let me
know how to proceed.

Emily

On Wed, Nov 20, 2024 at 7:34 PM Kanishka Narayan ***@***.***>
wrote:

> @XinyiEmilyZhang <https://github.com/XinyiEmilyZhang>, your name was
> recommended as a potential reviewer. Would you be interested in reviewing
> this package for the Journal of Open Source Software (JOSS)?
>
