### 2024-11-22 #7341 @XinyiEmilyZhang
@kanishkan91, thanks for adding me as a reviewer and providing all the necessary instructions.

I’m wondering if I should start the review now or wait for @LingfengLuo0510 to complete the first round as the review might have already started from the original reviewer side. However, I’d like to note that we’re running very short on time, as the original timeline suggests 4-6 weeks to complete the review process, though having three reviewers working simultaneously could complicate things for both the authors and editors.

Could you please advise on how best to proceed? Thanks so much.
