### 2024-11-28 7b90a04
added comments to functions in cox.py file
