### 2024-11-29 1fb825e
Update README.md
fixed a minor typo
