### 2024-12-05 #7341 @XinyiEmilyZhang
Thanks for building a comprehensive toolbox to provide survival analysis usage under the deep learning setting.

Comments for JOSS paper submission:

1. In the "Statement of Need" section, it would be beneficial to provide more detailed comments on the existing challenges in estimating survival parameters and performing full-body survival analysis within a **deep learning framework.**  By addressing these challenges, the importance of developing an "interpretable and easy-to-use" survival analysis toolbox, like the one presented in this paper, becomes more evident.
2. In the "Statement of Need" section, It would be great to include the need for having computationally efficient tools in deep survival analysis, as this would highlight one of the advantages of this framework.

Comments for TorchSurv Tutorial Website:

1. Please fix the typo under[ example ](https://opensource.nibr.com/torchsurv/_autosummary/torchsurv.loss.cox.html#torchsurv.loss.cox.neg_partial_log_likelihood)

`neg_partial_log_likelihood(log_hz, event, time, reduction = 'sum') # **sum** of log likelihoods across patients
`
`neg_partial_log_likelihood(log_hz, event, time, ties_method = "breslow")  # Dealing with ties **(Breslow)**
`
2. In the API documentation, it would be great to mention that an error message will appear when event ties exist while does not specify the parameter "ties_method".
3. Since most functions have been demonstrated with outputs using simulated toy datasets in the tutorial, could you add an example of the KM plot to showcase the visualization of your `plot_km()` function?

I will review the rest of the documentation and the code for next. Thanks!
