### 2024-12-05 4315011
fixed typos, added warning for ties
