### 2024-12-05 #7341 @tcoroller
thanks @XinyiEmilyZhang, we started the [PR](https://github.com/Novartis/torchsurv/pull/78) to answer your comments
