+++ docs/notebooks/helpers_momentum.py
@@ -162,6 +162,7 @@ class MNISTDataModule(L.LightningDataModule):
             batch_size=self.batch_size,
             num_workers=self.num_workers,
             persistent_workers=True,
+            shuffle=True,
         )
 
     def val_dataloader(self):
