+++ src/torchsurv/stats/kaplan_meier.py
@@ -4,16 +4,17 @@ from typing import Tuple
 
 import torch
 
-from ..tools import validate_inputs
+from torchsurv.tools import validate_data
 
 
+# @torch.jit.script
 class KaplanMeierEstimator:
     """Kaplan-Meier estimate of survival or censoring distribution for right-censored data :cite:p:`Kaplan1958`."""
 
     def __call__(
         self,
-        event: torch.tensor,
-        time: torch.tensor,
+        event: torch.Tensor,
+        time: torch.Tensor,
         censoring_dist: bool = False,
         check: bool = True,
     ):
@@ -62,7 +63,7 @@ class KaplanMeierEstimator:
 
         # Check input validity if required
         if check:
-            validate_inputs.validate_survival_data(event, time)
+            validate_data.validate_inputs(event, time)
 
         # Compute the counts of events, censorings, and the number at risk at each unique time
         uniq_times, n_events, n_at_risk, n_censored = self._compute_counts()
@@ -200,7 +201,7 @@ class KaplanMeierEstimator:
 
     def _compute_counts(
         self,
-    ) -> Tuple[torch.tensor, torch.tensor, torch.tensor, torch.tensor]:
+    ) -> Tuple[torch.Tensor, torch.Tensor, torch.Tensor, torch.Tensor]:
         """Compute the counts of events, censorings and risk set at ``time``.
 
         Returns: Tuple[torch.tensor, torch.tensor, torch.tensor, torch.tensor]
