+++ src/torchsurv/stats/kaplan_meier.py
@@ -7,7 +7,7 @@ import torch
 from torchsurv.tools import validate_data
 
 
-# @torch.jit.script
+@torch.jit.script
 class KaplanMeierEstimator:
     """Kaplan-Meier estimate of survival or censoring distribution for right-censored data :cite:p:`Kaplan1958`."""
 
