### 2024-12-10 bfad67b
cleanup, cox working
