### 2024-12-10 e020377
torch.compile runs without error and evaluates cox and weibull losses to the same value
