+++ tests/test_torch_compile.py
@@ -2,12 +2,6 @@
 
 Tests for torch.compile
 
-Note: conda install conda-forge::gxx
-
-TODO: 
-    - test conda install conda-forge::cxx-compiler
-    - check torchtriton for GPU
-    
 References:
     - https://pytorch.org/tutorials/intermediate/torch_compile_tutorial.html
     - https://github.com/pytorch/pytorch/issues/122094
@@ -28,6 +22,7 @@ from torchsurv.loss.weibull import neg_log_likelihood as weibull
 # set seed for reproducibility
 torch.manual_seed(42)
 
+
 class TestTorchCompile(unittest.TestCase):
     """
     Tests using torch.compile with cox
@@ -37,7 +32,7 @@ class TestTorchCompile(unittest.TestCase):
         """
         whether the compiled version of cox evaluates to the same value
         """
-        
+
         # random data and parameters
         N = 32
         log_hz = torch.randn(N)
@@ -50,16 +45,15 @@ class TestTorchCompile(unittest.TestCase):
         loss_cox = cox(log_hz, event, time)
         loss_ccox = ccox(log_hz, event, time)
 
-        self.assertTrue(np.isclose(loss_cox.numpy(),
-                                   loss_ccox.numpy(),
-                                   rtol=1e-3,
-                                   atol=1e-3))
+        self.assertTrue(
+            np.isclose(loss_cox.numpy(), loss_ccox.numpy(), rtol=1e-3, atol=1e-3)
+        )
 
     def test_weibull_equivalence(self):
         """
         whether the compiled version of weibull evaluates to the same value
         """
-        
+
         # random data and parameters
         N = 32
         log_hz = torch.randn(N)
@@ -72,16 +66,13 @@ class TestTorchCompile(unittest.TestCase):
         loss_weibull = weibull(log_hz, event, time)
         loss_cweibull = cweibull(log_hz, event, time)
 
-        self.assertTrue(np.isclose(loss_weibull.numpy(),
-                                   loss_cweibull.numpy(),
-                                   rtol=1e-3,
-                                   atol=1e-3))
+        self.assertTrue(
+            np.isclose(
+                loss_weibull.numpy(), loss_cweibull.numpy(), rtol=1e-3, atol=1e-3
+            )
+        )
+
 
 if __name__ == "__main__":
-    
-    unittest.main()
 
-    
-    
-    
-    
+    unittest.main()
