### 2024-12-11 ffd6920
added comments to functions in cox.py file (#76)
* added comments to functions in cox.py file

* linting

* missing escape

---------

Co-authored-by: corolth1 <thibaud.coroller@novartis.com>
