### 2024-12-11 093ce0b
working doctest, cleanup and ready to check pytest on CI
