### 2024-12-11 a505bfc
JOSS #3  (#78)
* fixed typos, added warning for ties

* removed lines, added plots

* cleanup

* Address reviews on the Statement of Need Section

---------

Co-authored-by: melodiemonod <monod.melodie@gmail.com>
