### 2024-12-11 c72c5d5
used torch instead of numpy
