### 2024-12-11 fbfdb99
moving new files over to otebooks folder, added comments and changed the log partial likelihood function
