### 2024-12-12 #7341 @tcoroller
Hello @WeakCha , I have completed your suggestions in the text (thanks again!) but I am unable to replicate your error (neither are my other colleagues), so unless you can provide an minimal example I would suggest keeping this one aside and move it as a git issue to our package rather than a review request, if you agree.

Thanks!
