### 2024-12-12 #7341 @kanishkan91
@tcoroller @WeakCha Let's move this issue to git. It seems related to python versions.

@tcoroller correct me if im wrong, but regarding @WeakCha 's other comment, the difference between the c indexes is not a p value, right? You would have to convert the cindex values to a z value before converting? It seems this is just a differences in the indices. If so I would just say that and remove the p value terminology. Let me know what you think.
