### 2024-12-12 #7341 @melodiemonod
Dear @kanishkan91, thank you for your comment. To compare the two c-indexes, we indeed use a statistical test that gives a z-value and a p-value (more details and relevant reference in our documentation here https://opensource.nibr.com/torchsurv/_autosummary/torchsurv.metrics.cindex.html#torchsurv.metrics.cindex.ConcordanceIndex.compare). This is mentioned in our paper as follows:
>  Lastly, functionalities including the confidence interval, one-sample hypothesis test to determine whether the metric is better than that of a random predictor, and two-sample hypothesis test to compare two evaluation metrics between different models are implemented.
