### 2024-12-13 #7341 @kanishkan91
@melodiemonod @tcoroller Thanks. Though I still maintain that you should take out the p value statement. It does not seem statistically significant even though the value is lower. I think it should be enough to say that cindex values are lower on that line without reference to the p value.
