### 2024-12-13 #7341 @tcoroller
@kanishkan91 I fully agree with you, this is a mistake from my part.

I had to modify the training (sped up using only 5 batches instead of 200) to implement suggestion from @WeakCha. I just [fixed](https://github.com/Novartis/torchsurv/pull/84) it to reflect our original submission, and the conclusion holds now at the end of the [notebook](https://github.com/Novartis/torchsurv/blob/main/docs/notebooks/momentum.ipynb).

Thank you both for pointing this out!
