### 2024-12-13 0d3905b
remove unused code, added comments, simulation in lifelines
