### 2024-12-16 #7341 @tcoroller
> @kanishkan91 Thanks a lot!
>
>
>
> @tcoroller I found you reran the experiments, however, I have to say the conclusion was still incorrect.
>
>
>
> <img width="991" alt="image" src="https://github.com/user-attachments/assets/b9e0e549-03e5-4e25-b53e-1d07f3a5779b" />
>
>
>
> The results shown in your tutorial are also different from those you put in the review thread. Your conclusion in the screenshot is correct, but not reflecting on the tutorial website.
>
>
>
> Could you please double check to see what might go wrong?
>
>
>
> Thanks!

Hello! I seemed that the GitHub-page may not reflect the changes on the notebook, but does reflects other changes (like the README.md) - this is strange but an easy fix.

Let us know if there is anything else we can do to complete the review. Thanks a lot!
