### 2024-12-16 #7341 @tcoroller
> @kanishkan91 Thanks a lot!
>
> @tcoroller I found you reran the experiments, however, I have to say the conclusion was still incorrect.
>
> <img alt="image" width="991" src="https://private-user-images.githubusercontent.com/48302151/395815235-b9e0e549-03e5-4e25-b53e-1d07f3a5779b.png?jwt=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJnaXRodWIuY29tIiwiYXVkIjoicmF3LmdpdGh1YnVzZXJjb250ZW50LmNvbSIsImtleSI6ImtleTUiLCJleHAiOjE3MzQzNTg5ODUsIm5iZiI6MTczNDM1ODY4NSwicGF0aCI6Ii80ODMwMjE1MS8zOTU4MTUyMzUtYjllMGU1NDktMDNlNS00ZTI1LWI1M2UtMWQwN2YzYTU3NzliLnBuZz9YLUFtei1BbGdvcml0aG09QVdTNC1ITUFDLVNIQTI1NiZYLUFtei1DcmVkZW50aWFsPUFLSUFWQ09EWUxTQTUzUFFLNFpBJTJGMjAyNDEyMTYlMkZ1cy1lYXN0LTElMkZzMyUyRmF3czRfcmVxdWVzdCZYLUFtei1EYXRlPTIwMjQxMjE2VDE0MTgwNVomWC1BbXotRXhwaXJlcz0zMDAmWC1BbXotU2lnbmF0dXJlPThhODk1ZDA0MThmNDg2M2Y5YzZiODA4Y2VhMWI3NTA2ODk3MmIxNWUzY2VlODZmZWI1ZTk0NDUwODliMGZkMzgmWC1BbXotU2lnbmVkSGVhZGVycz1ob3N0In0.AXNJYYQezQj79ssr0vnstjiTyqnNfq3ObRhUfLmjuGg">
> The results shown in your tutorial are also different from those you put in the review thread. Your conclusion in the screenshot is correct, but not reflecting on the tutorial website.
>
> Could you please double check to see what might go wrong?
>
> Thanks!

Thanks for pointing this out, it seems that the [github pages notebook](https://opensource.nibr.com/torchsurv/notebooks/momentum.html) is not reflecting the [latest  notebook version](https://github.com/Novartis/torchsurv/blob/main/docs/notebooks/momentum.ipynb) state values (cell outputs).
We will fix the documentation build ASAP, which should close this review item.

Please let us know if there is anything else can do to complete the review!

Thank you
