### 2024-12-16 #7341 @melodiemonod
Dear @kanishkan91 , thank you for your message. We will make the GitHub release associated with the JOSS reviews tonight and will share the package version and DOI tomorrow.
