### 2024-12-16 #7341 @tcoroller
We just released the new version on PYPI (conda PR is pending). you can find the release tag here:
https://github.com/Novartis/torchsurv/releases/tag/v0.1.4

Thanks!
