### 2024-12-17 #7341 @kanishkan91
@tcoroller I will need a valid DOI for the archive. Is this linked to Zenodo or any similar archiving website by chance? Else I would recommend just connecting your GitHub to zenodo and getting a DOI from there.
