### 2024-12-17 dafdb40
small changes to both the notebook and loss files
