### 2024-12-18 a141f8d
loss function works, added function documentation
