+++ docs/notebooks/loss_time_covariates.py
@@ -5,23 +5,16 @@ import torch
 
 
 def neg_partial_time_log_likelihood(
-    log_hz: torch.Tensor,
-    event: torch.Tensor,
-    time: torch.Tensor,
-    ties_method: str = "efron",
+    log_hz: torch.Tensor, #Txnxp torch tensor, n is batch size, T number of time points, p is number of different covariates over time
+    time: torch.Tensor, #n length vector, time at which someone experiences event
+    events: torch.Tensor, #n length vector, boolean, true or false to determine if someone had an event
     reduction: str = "mean"
 ) -> torch.Tensor:
     '''
     needs further work
     '''
-    # sort data by time-to-event or censoring
-    time_sorted, idx = torch.sort(time)
-    log_hz_sorted = log_hz[idx]
-    event_sorted = event[idx]
-    time_unique = torch.unique(time_sorted)  # time-to-event or censoring without ties
-
     # only consider theta at tiem of
-    pll = _partial_likelihood_time_cox(log_hz_sorted, event_sorted)
+    pll = _partial_likelihood_time_cox(log_hz, time, events)
    
     # Negative partial log likelihood
     pll = torch.neg(pll)
@@ -38,47 +31,82 @@ def neg_partial_time_log_likelihood(
     return loss
 
 def _partial_likelihood_time_cox(
-    log_hz: torch.Tensor, #nxTxp torch tensor, n is batch size, T number of time points, p is number of different covariates over time
-    event: torch.Tensor, #n length vector, boolean, true or false to determine if someone had an event
+    log_hz: torch.Tensor, #Txnxp torch tensor, n is batch size, T number of time points, p is number of different covariates over time
     time: torch.Tensor, #n length vector, time at which someone experiences event
+    events: torch.Tensor, #n length vector, boolean, true or false to determine if someone had an event
+    
 ) -> torch.Tensor:
-    """Calculate the partial log likelihood for the Cox proportional hazards model
+    """
+    Calculate the partial log likelihood for the Cox proportional hazards model
     with time-varying covariates and in the absence of ties in event time.
 
-    For time-varying covariates, the haard ratio is no longer assumed to be constant, 
-    but the partial log likelihood only cares about the covariate value at time of death.
+    Args:
+        log_hz (torch.Tensor, float):
+            Log relative hazard of dimension T x n_samples x P.
+            T is the time series dimension, P is the number of parameters observed over time.
+        event (torch.Tensor, bool):
+            Event indicator of length n_samples (= True if event occured).
+        time (torch.Tensor):
+            Time-to-event or censoring of length n_samples.
 
-    Hence, despite taking in a whole vector of stuff, we only take the last value 
-    into consideration for the partial log likelihood.
+    Returns:
+        (torch.tensor, float):
+            Vector of the partial log likelihood, length n_samples.
 
-    Requirements we want:
-    - time vector must somehow correspond to the T dimension in the log_hz tensor, i.e. for those who experience an event,
+    Note:
+        For each subject :math:`i \in \{1, \cdots, N\}`, denote :math:`\tau^*_i` as the survival time and :math:`C_i` as the
+        censoring time. Survival data consist of the event indicator, :math:`\delta_i=1(\tau^*_i\leq C_i)`
+        (argument ``event``) and the time-to-event or censoring, :math:`\tau_i = \min(\{ \tau^*_i,D_i \})`
+        (argument ``time``).
+
+        Consider some covariate :math:`Z(t)` with covariate history denoted as :math:`H_Z` and a general form of the cox proportional hazards model:
+            .. math::
+
+            \log \lambda_i (t|H_Z) = lambda_0(t) \theta(Z(t))
+        
[…]
