### 2024-12-18 Novartis/torchsurv#85 @melodiemonod
FInal version JOSS paper
- [x] Correct final typos
- [x] Add Peter's ORCID
- [x] Remove Figure 2 which is not cited
