+++ _sources/AUTHORS.md
@@ -1,0 +1,10 @@
+Contributors
+============
+
+* Thibaud Coroller <thibaud.coroller@novartis.com> `(creator, maintainer)`
+* Mélodie Monod <melodie.monod@novartis.com> `(creator, maintainer)`
+* Peter Krusche <peter.krusche@novartis.com> `(author, maintainer)`
+* Qian Cao <qian.cao@fda.hhs.gov> `(author, maintainer)`
+* David Ohlssen <david.ohlssen@novartis.com> `(contributor)`
+* Berkman Sahiner <berkman.sahiner@fda.hhs.gov> `(contributor)`
+* Nicholas Petrick <nicholas.petrick@fda.hhs.gov> `(contributor)`
