### 2024-12-18 #7341 @melodiemonod
Dear @kanishkan91 ,
Thank you for your patience. We have made the final modifs to the paper (correction of minor typos and addition of the ORCID from one author -- you can check the [issue](https://github.com/Novartis/torchsurv/issues/85) and [PR](https://github.com/Novartis/torchsurv/pull/86) related to this most updated paper version to see the edits). Following this, we have made the Zenodo archive.

Version: v0.1.4
Zenodo doi: https://doi.org/10.5281/zenodo.14517267
URL: https://zenodo.org/records/14517267
