### 2024-12-20 #7341 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
