### 2024-12-20 #7341 @kanishkan91
@tcoroller , @melodiemonod  I have recommended this for acceptance now. I will be reading through the paper for typos etc. shortly. The AEiC in this submission track will review shortly and if all goes well this will go live soon! Big thank you to @rich2355 and @XinyiEmilyZhang for reviewing! JOSS is volunteer run and relies heavily on researchers such as yourself.
