### 2024-12-20 #7341 @kanishkan91
@melodiemonod @tcoroller - I guess one last minor comment from my side. The paper is a bit too long. We generally only recommend 3-4 pages of text for a software paper. I recommend removing the "Comprehensive example"  starting page 4. This is probably covered in your documentation.
