### 2024-12-20 #7341 @melodiemonod
Dear @kanishkan91, thank you for your feedback. We are open to removing the example, provided that @rich2355, who originally suggested it, agrees. Please let us know your thoughts so we can proceed accordingly.
