+++ _sources/index.rst
@@ -1,0 +1,38 @@
+Welcome to TorchSurv's documentation!
+=====================================
+
+.. include:: README-copy.md
+    :parser: myst_parser.sphinx_
+
+.. toctree::
+    :maxdepth: 2
+    :caption: API:
+
+    loss
+    metrics
+    stats
+
+.. toctree::
+    :maxdepth: 2
+    :caption: Tutorials:
+
+    notebooks/survival
+    notebooks/introduction
+    notebooks/momentum
+
+.. toctree::
+    :maxdepth: 2
+    :caption: Other:
+
+    CHANGELOG
+    AUTHORS
+    devnotes
+    benchmarks
+    
+    
+Indices and tables
+==================
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
