### 2024-12-20 #7341 @crvernon
:wave: @melodiemonod - I just need you to address the following before I move to accept this for publication:

In the archive:

- [ ] The names listed in your Zenodo archive do not match the names listed in your paper.  Please edit the Zenodo metadata to ensure that the names and their order match what is in the paper exactly.

- [ ] LINE 93: "pytorch" should be written as "PyTorch".  You can maintain capitalization in your bib file by using curly brackets around the characters you wish to maintain formatting of.
- [ ] LINE 101: the "w" in "weibull" should be capitalized
- [ ] LINE 107: the "p" in  "python" should be capitalized
- [ ] LINE 109: the "c" in "cox's" should be capitalized
- [ ] LINE 126: the "c" in "cox" should be capitalized
- [ ] LINE 130 the "c" in "cox" should be capitalized
- [ ] LINE 144 the "c" in "c-statistic" should be capitalized

After you make these changes let me know.  Thanks.
